backed_block.c: support img file larger than 4GB
change the len member of backed_block to int64_t
from unsigned int type, so that we can support image file
larger than 4GB size.

+++ libsparse/backed_block.c
@@ -25,7 +25,7 @@
 
 struct backed_block {
 	unsigned int block;
-	unsigned int len;
+	int64_t len;
 	enum backed_block_type type;
 	union {
 		struct {
